# app/openapi.yml
openapi: "3.0.0"
servers:
  - description:  Test
    url: /
info:
  description: This is a simple API
  version: "1.0.0"
  title: Test API
paths:
...
  /dealer/{dealer_id}:
    get:
      tags:
        - Dealer
      summary: Get dealer by dealer_id 
      operationId: app.views.dealerView.get_dealer_by_id
      description: |
        Get dealer by dealer_id 
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDealerResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  
    patch:
      tags:
        - Dealer
      summary: Update/Patch Dealer by dealer id
      operationId: app.views.dealerView.update_dealer
      description: |
        Updates a single record based on dealer_id 
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
      requestBody:
        description: Updates a dealer inventory record
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealerPatch'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Dealer
      summary: Deletes Dealer by dealer id
      operationId: app.views.dealerView.delete_dealer
      description: |
        Deletes a single dealer  record 
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /employee:
    post:
      tags:
        - Employee
      summary: Add a new Employee
      operationId: app.views.employeeView.add_employee
      description: |
        Add a new Employee
      requestBody:
        description: Creates a new Employee 
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags:
        - Employee
      summary: List all employees
      operationId: app.views.employeeView.get_employee
      description: |
        List all employees
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllEmployeeResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /employee/{employee_id}:
    get:
      tags:
        - Employee
      summary: Get employee by employee_id
      operationId: app.views.employeeView.get_employee_by_id
      description: |
        Get employee by employee_id
      parameters:
        - name: employee_id
          in: path 
          description: Employee ID
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEmployeeResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  
    patch:
      tags:
        - Employee
      summary: Update/Patch Employee by employee id
      operationId: app.views.employeeView.update_employee
      description: |
        Updates a single record based on employee_id 
      parameters:
        - name: employee_id
          in: path 
          description: Employee ID
          required: true 
          schema:
            type: integer 
            example: 1
      requestBody:
        description: Updates a employee record
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeePatch'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Employee
      summary: Deletes Employee by Employee ID
      operationId: app.views.employeeView.delete_employee
      description: |
        Deletes a single employee  record 
      parameters:
        - name: employee_id
          in: path 
          description: Employee ID
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /employee/dealer/{dealer_id}:
    get:
      tags:
        - Employee
      summary: Get all employees by dealer id
      operationId: app.views.employeeView.get_employee_by_dealer
      description: |
        Get all employees working in a dealership
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllEmployeeResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /sale_item:
    post:
      tags:
        - Sale_Item
      summary: Add a new Sale Line Item
      operationId: app.views.saleLineItemView.add_sale_line_item
      description: |
        Add a new Invoice Line Item with cart details 
      requestBody:
        description: Creates a new Sale Line Item  
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleLineItem'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /sale_item/{invoice_id}:
    get:
      tags:
        - Sale_Item
      summary: List all sales line items for a given invoice 
      operationId: app.views.saleLineItemView.get_sale_item
      description: |
        List all sale line items for a given invoice id 
      parameters:
        - name: invoice_id
          in: path 
          description: Invoice ID
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllSaleLineItemResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /sale_item/{invoice_line_item_id}:
    patch:
      tags:
        - Sale_Item
      summary: Update/Patch Sale line item  by invoice_line_item_id  
      operationId: app.views.saleLineItemView.update_sale_item
      description: |
        Update/Patch Sale line item  by invoice_line_item_id  
      parameters:
        - name: invoice_line_item_id
          in: path 
          description: invoice_line_item_id
          required: true 
          schema:
            type: integer 
            example: 1
      requestBody:
        description: Updates a sale line item record
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleLineItemPatch'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Sale_Item
      summary: Deletes Sale line item by invoice_line_item_id
      operationId: app.views.saleLineItemView.delete_sale_item
      description: |
        Deletes a single sale  record 
      parameters:
        - name: invoice_line_item_id
          in: path 
          description: invoice_line_item_id
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /state_tax/{state_tax_id}:
    get:
      tags:
        - State_Tax
      summary: Get tax information by state_tax_id
      operationId: app.views.stateTaxView.get_state_by_id
      description: |
        Get tax information by state_tax_id
      parameters:
        - name: state_tax_id
          in: path 
          description: state_tax_id
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetStateTaxResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        
    patch:
      tags:
        - State_Tax
      summary: Update/Patch State Tax  by state_tax_id
      operationId: app.views.stateTaxView.update_state_tax
      description: |
        Update/Patch Sale line item  by state_tax_id
      parameters:
        - name: state_tax_id
          in: path 
          description: state_tax_id
          required: true 
          schema:
            type: integer 
            example: 1
      requestBody:
        description: Update/Patch State Tax  by state_tax_id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateTaxPatch'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - State_Tax
      summary: Deletes state tax by state_tax_id
      operationId: app.views.stateTaxView.delete_state_tax
      description: |
        Deletes a single state tax  record 
      parameters:
        - name: state_tax_id
          in: path 
          description: state_tax_id
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
...
    DealerObject:
      type: object
      properties:
        dealer_id:
          type: integer
          example: 1 
        name:
          type: string
          example: Cartec Honda
        web_url:
          type: string 
          example: "cartechonda@example.com"
        street_address:
          type: string 
          example: Abc Street, CA-94301
        postal_code:
          type: string 
          example: 94301
        city:
          type: string 
          example: Palo Alto
        state:
          type: string 
          example: california
        created_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT  
        modified_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT
    GetDealerResponse:
      type: object 
      properties:
        msg:
          $ref: '#/components/schemas/DealerObject'
        status:
          type: string 
          example: success
    DealerPatch:
      type: object
      properties:
        name:
          type: string
          example: Cartec Honda
        web_url:
          type: string 
          example: "cartechonda@example.com"
        street_address:
          type: string 
          example: Abc Street, CA-94301
        postal_code:
          type: string 
          example: 94301
        city:
          type: string 
          example: Palo Alto
        state:
          type: string 
          example: california
    Employee:
      required: 
        - first_name
        - last_name
        - dealer_id 
      type: object
      properties:
        first_name:
          type: string
          example: John
        last_name:
          type: string 
          example: Doe
        dealer_id:
          type: integer 
          example: 1
    GetAllEmployeeResponse:
      type: object 
      properties:
        msg:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeObject'
        status:
          type: string 
    EmployeeObject:
      type: object
      properties:
        first_name:
          type: string
          example: John
        last_name:
          type: string 
          example: Doe
        dealer_id:
          type: integer 
          example: 1
        employee_id:
          type: integer
          example: 1 
        created_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT  
        modified_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT
    GetEmployeeResponse:
      type: object 
      properties:
        msg:
          $ref: '#/components/schemas/EmployeeObject'
        status:
          type: string 
          example: success
    EmployeePatch:
      type: object
      properties:
        first_name:
          type: string
          example: John
        last_name:
          type: string 
          example: Doe
        dealer_id:
          type: integer 
          example: 1
...
    SaleLineItem:
      required: 
        - invoice_id
        - dealer_id
        - cart
      type: object
      properties:
        invoice_id:
          type: integer
          example: 1
        dealer_id:
          type: integer 
          example: 1
        cart:
          type: array
          items:
            $ref: '#/components/schemas/SaleLineCartObject'
    SaleLineCartObject:
      required:
        - car_serial_no
      type: object 
      properties: 
        car_serial_no:
          type: string 
          example: C-1
        quantity:
          type: integer
          example: 2 
        discount_percentage:
          type: integer 
          example: 10 
    GetAllSaleLineItemResponse:
      type: object 
      properties:
        msg:
          type: array
          items:
            $ref: '#/components/schemas/SaleLineItemObject'
        status:
          type: string 
    SaleLineItemObject:
      type: object 
      properties:
        car_serial_no:
          type: string 
          example: C-1
        quantity:
          type: integer
          example: 2 
        discount_percentage:
          type: integer 
          example: 10 
        car_net_price:
          type: number 
          example: 151010.0
        car_unit_price:
          type: number 
          example: 151010.0
        invoice_id:
          type: integer
          example: 1 
        invoice_line_item_id:
          type: integer
          example: 1 
        item_tax:
          type: number 
          example: 10.0
        created_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT  
        modified_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT
    SaleLineItemPatch:
      type: object
      properties:
        car_serial_no:
          type: string 
          example: C-1
        quantity:
          type: integer
          example: 2 
        discount_percentage:
          type: integer 
          example: 10 
        car_net_price:
          type: number 
          example: 151010.0
        car_unit_price:
          type: number 
          example: 151010.0
        invoice_id:
          type: integer
          example: 1 
        item_tax:
          type: number 
          example: 10.0
...
    StateTaxObject:
      type: object
      properties:
        state_tax_id:
          type: integer
          example: 1 
        state:
          type: string 
          example: california
        tax: 
          type: number 
          example: 10 
        created_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT  
        modified_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT
    GetStateTaxResponse:
      type: object 
      properties:
        msg:
          $ref: '#/components/schemas/StateTaxObject'
        status:
          type: string 
          example: success
    StateTaxPatch:
      type: object
      properties:
        tax: 
          type: number 
          example: 10 
    ApiResponse:
        type: object
        properties:
          msg:
            type: string
          status:
            type: string 